Update test infrastructure
