--- test/src/word-test-taskpane.ts
@@ -1,40 +1,0 @@
-import { pingTestServer, sendTestResults } from "office-addin-test-helpers";
-import * as word from "../../src/taskpane/app/word.app.component";
-import * as testHelpers from "./test-helpers";
-const port: number = 4201;
-let testValues: any = [];
-
-Office.onReady(async(info) => {
-    if (info.host === Office.HostType.Word) {
-        const testServerResponse: object = await pingTestServer(port);
-        if (testServerResponse["status"] == 200) {
-            await runTest();
-        }  
-    }
-});
-
-export async function runTest() {
-    return new Promise<void>(async (resolve, reject) => {
-        try {
-            // Execute taskpane code
-            const wordComponent = new word.default();
-            await wordComponent.run();
-            await testHelpers.sleep(2000);
-
-            // Get output of executed taskpane code
-            Word.run(async (context) => {
-                var firstParagraph = context.document.body.paragraphs.getFirst();
-                firstParagraph.load("text");
-                await context.sync();
-                await testHelpers.sleep(2000);
-
-                testHelpers.addTestResult(testValues, "output-message", firstParagraph.text, "Hello World");
-                await sendTestResults(testValues, port);
-                testValues.pop();
-                resolve();
-            });
-        } catch {
-            reject();
-        }
-    });
-}
